+++ sentry/src/test/java/io/sentry/SentryTest.kt
@@ -1,5 +1,7 @@
 package io.sentry
 
+import io.sentry.cache.EnvelopeCache
+import io.sentry.cache.IEnvelopeCache
 import io.sentry.protocol.SentryId
 import org.junit.rules.TemporaryFolder
 import org.mockito.kotlin.any
@@ -264,6 +266,39 @@ class SentryTest {
         }
 
         verify(logger).log(eq(SentryLevel.ERROR), any(), eq(initException))
+    }
+
+    @Test
+    fun `overrides envelope cache if it's not set`() {
+        var sentryOptions: SentryOptions? = null
+
+        Sentry.init {
+            it.dsn = dsn
+            it.cacheDirPath = getTempPath()
+            sentryOptions = it
+        }
+
+        assertTrue { sentryOptions!!.envelopeDiskCache is EnvelopeCache }
+    }
+
+    @Test
+    fun `does not override envelope cache if it's already set`() {
+        var sentryOptions: SentryOptions? = null
+
+        Sentry.init {
+            it.dsn = dsn
+            it.cacheDirPath = getTempPath()
+            it.setEnvelopeDiskCache(CustomEnvelopCache())
+            sentryOptions = it
+        }
+
+        assertTrue { sentryOptions!!.envelopeDiskCache is CustomEnvelopCache }
+    }
+
+    private class CustomEnvelopCache : IEnvelopeCache {
+        override fun iterator(): MutableIterator<SentryEnvelope> = TODO()
+        override fun store(envelope: SentryEnvelope, hint: Hint) = Unit
+        override fun discard(envelope: SentryEnvelope) = Unit
     }
 
     private fun getTempPath(): String {